Issue #459: refactor EN locale
